% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_nbm.R
\name{dl_nbm}
\alias{dl_nbm}
\title{Download NBM data}
\usage{
dl_nbm(
  path_to_dl = "~/data_swamp",
  release_date = "June 30, 2023",
  data_type = "Fixed Broadband",
  data_category = "Nationwide",
  user_agent = the$user_agent,
  ...
)
}
\arguments{
\item{path_to_dl}{a string by default "~/data_swamp"}

\item{release_date}{a string can be "December 31, 2023" or "June 30, 2023"}

\item{data_type}{a string "Fixed Broadband"}

\item{data_category}{a string "Nationwide"}

\item{user_agent}{a string set up by default}

\item{...}{additional parameters for download.file()}
}
\value{
A lot of zipped file
}
\description{
Just a draft of a function that download all NBM data related to CORI works
It motsly works in my setup (download in ~/data_swamp)
}
\examples{
\dontrun{
system("mkdir -p  ~/data_swamp")
dl_nbm(release_date = "June 30, 2023")
}
}
