% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download file function (replacement for download.file)}
\usage{
download_file(remote_file_url, local_file_path)
}
\arguments{
\item{remote_file_url}{URL to download file from}

\item{local_file_path}{Local path to save file to}
}
\value{
path to local file
}
\description{
Download file function (replacement for download.file)
}
\keyword{internal}
