% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc_to_parquet.R
\name{fcc_to_parquet}
\alias{fcc_to_parquet}
\title{Convert CSVs to parquet for NBM data}
\usage{
fcc_to_parquet(
  parquet_name,
  src_directory = "~/data_swamp/",
  part_one = "state_abbr",
  part_two = "technology"
)
}
\arguments{
\item{parquet_name}{a path were the parquet files should be stored}

\item{src_directory}{a directory containing only the csv needed}

\item{part_one}{a string defining the first column to partition the parquet file}

\item{part_two}{a string defining the second column to partition the parquet file}
}
\value{
0 if the execute went well and write a parquet file
}
\description{
Draft version of converting a lot of FCC NBM CSVs into a parquet file system
}
\examples{
\dontrun{
system("unzip ~/data_swamp/\*.zip")
fcc_to_parquet("june23")
}
}
