% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f477.R
\name{get_f477}
\alias{get_f477}
\title{Load part of f477 estimates data from s3 bucket}
\usage{
get_f477(us_states)
}
\arguments{
\item{us_states}{a directory containing only the csv needed}
}
\value{
a data frame
}
\description{
Source data: FCC Form 477
}
\examples{
\dontrun{
NC <- get_f477(us_states = "NC")
}
}
