% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc_staff.R
\name{get_fcc_staff}
\alias{get_fcc_staff}
\title{Load part of fCC Staff estimates data from s3 bucket}
\usage{
get_fcc_staff(year, us_states)
}
\arguments{
\item{year}{numeric a year of f477}

\item{us_states}{a directory containing only the csv needed}
}
\value{
a data frame
}
\description{
Load part of fCC Staff estimates data from s3 bucket
}
\examples{
\dontrun{
vt_2017 <- get_fcc_staff(year = 2017, us_states = "VT")
}
}
