% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_available.R
\name{get_nbm_available}
\alias{get_nbm_available}
\title{Get release available in FCC NBM}
\usage{
get_nbm_available(
  get_data_url =
    "https://broadbandmap.fcc.gov/nbm/map/api/national_map_process/nbm_get_data_download/"
)
}
\arguments{
\item{get_data_url}{a string providing NBM filing API. Default is "https://broadbandmap.fcc.gov/nbm/map/api/national_map_process/nbm_get_data_download/"}
}
\value{
A data frame.
}
\description{
Get release available in FCC NBM
}
\examples{
nbm_data <- get_nbm_available()
}
