% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_county_raw.R
\name{get_nbm_county_raw}
\alias{get_nbm_county_raw}
\title{Load all BSL from NBM dataset for a specific county}
\usage{
get_nbm_county_raw(geoid_co, frn = "all", release = "2024-12-01")
}
\arguments{
\item{geoid_co}{a string matching a GEOID for a county}

\item{frn}{a string of 10 numbers matching FCC's FRN, default is "all"}

\item{release}{a date, set by default to be '2024-12-01'}
}
\value{
a data frame
}
\description{
Attention: Depending on the county it can take some times.
}
\details{
Get all the BSL from NBM dataset for a specific county specified by it's geoid.
A row in this data represent a service per location_id.
By default the function will return all the ISP but a specific FRN and the last release.
(10 number strings, ie "0007435902") can also be used to be more specific.

Data Source: FCC Broadband Data Collection
}
\examples{
\dontrun{
 guilford_cty <- get_nbm_county_raw(geoid_co = "37081")
}
}
