% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_release.R
\name{get_nbm_release}
\alias{get_nbm_release}
\title{Get a list of release available in FCC NBM}
\usage{
get_nbm_release(
  filing_url = "https://broadbandmap.fcc.gov/nbm/map/api/published/filing"
)
}
\arguments{
\item{filing_url}{a string providing NBM filing API. Default is "https://broadbandmap.fcc.gov/nbm/map/api/published/filing"}
}
\value{
A data frame.
}
\description{
Get a list of release available in FCC NBM
}
\examples{
nbm <- get_nbm_release()
}
