% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbm_csv_to_parquet.R
\name{nbm_csv_to_parquet}
\alias{nbm_csv_to_parquet}
\title{Convert CSVs to parquet for NBM data}
\usage{
nbm_csv_to_parquet(
  parquet_name,
  src_directory = "~/data_swamp/",
  part_one = "state_abbr",
  part_two = "technology"
)
}
\arguments{
\item{parquet_name}{a path were the parquet files should be stored}

\item{src_directory}{a directory containing only the csv needed}

\item{part_one}{a string defining the first column to partition the parquet file}

\item{part_two}{a string defining the second column to partition the parquet file}
}
\value{
0 if the execute went well and write a parquet file
}
\description{
Draft version of converting a lot of FCC NBM CSVs into a parquet file system
DuckDB is quite greedy on your system (but fast)
}
\keyword{internal}
