% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbm_dictionary.R
\docType{data}
\name{nbm_dictionary}
\alias{nbm_dictionary}
\title{Dictionary of FCC NBM fields}
\format{
\subsection{\code{nbm}}{

A data frame with XXXX rows and Y columns:
\describe{
\item{data_type}{data type}
\item{data_category}{data category}
\item{field}{field}
\item{description}{description}
}
}
}
\source{
\url{https://us-fcc.app.box.com/v/bdc-data-downloads-output}
}
\usage{
nbm_dictionary
}
\description{
Listing FCC specifications for every columns available in FCC
}
\keyword{datasets}
