% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frn.R
\name{check_frn}
\alias{check_frn}
\title{Get info on 1 frn}
\usage{
check_frn(frn)
}
\arguments{
\item{frn}{}
}
\value{
a table with info on FRN
}
\description{
Get info on 1 frn
}
\examples{
check_frn(8181448)
}
