% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frn.R
\name{check_frn}
\alias{check_frn}
\title{Get info on one frn}
\usage{
check_frn(frn)
}
\arguments{
\item{frn}{a FCC FRN, it can be a number or a string}
}
\value{
a table with info on FRN
}
\description{
Get info on one frn
}
\examples{
check_frn(81814488181448)
}
