% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc_dictionary.R
\docType{data}
\name{fcc_dictionary}
\alias{fcc_dictionary}
\title{Dictionary for our datasets}
\format{
\subsection{\code{fcc_dictionary}}{

A data frame with 15 rows and 5 columns:
\describe{
\item{dataset}{name of the dataset}
\item{var_name}{name of a field/columns}
\item{var_type}{Data type}
\item{var_description}{Description provided by FCC}
\item{var_example}{One illustration}
}
}
}
\source{
\url{https://www.fcc.gov/general/explanation-broadband-deployment-data}
}
\usage{
fcc_dictionary
}
\description{
Currently just provide a dictionary for Form 477
}
\keyword{datasets}
