% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc_provider.R
\docType{data}
\name{fcc_provider}
\alias{fcc_provider}
\title{FCC providers list}
\format{
\subsection{\code{fcc_provider}}{

A data frame with 4,341 rows and 5 columns:
\describe{
\item{Provider.Name}{Provider name}
\item{Affiliation}{Affiliation}
\item{Operation.Type}{ILEC or not}
\item{FRN}{FCC Registration Number, number of the entity that submited the data}
\item{Provider.ID}{Provider ID}
}
}
}
\source{
\url{https://us-fcc.app.box.com/v/bdcprovideridtable}
}
\usage{
fcc_provider
}
\description{
FCC list of preoviders downloaded in May 2024
}
\keyword{datasets}
