% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcc_dictionary.R
\name{get_fcc_dictionary}
\alias{get_fcc_dictionary}
\title{Display FCC variables and associated descriptions}
\usage{
get_fcc_dictionary(dataset = "all")
}
\arguments{
\item{dataset}{a string matching a dataset, default is "all"}
}
\value{
a data frame
}
\description{
Return dictionary for a all datasets. Available dataset are "f477", "nbm_raw" and "nbm_block".
}
\examples{
get_fcc_dictionary("nbm_block")
}
