% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcc_dictionary.R
\name{get_fcc_dictionary}
\alias{get_fcc_dictionary}
\title{Display FCC variable and descriptions}
\usage{
get_fcc_dictionary(dataset)
}
\arguments{
\item{dataset}{a string matching a dataset}
}
\value{
a data frame
}
\description{
Display FCC variable and descriptions
}
\examples{
get_fcc_dictionary("nbm_block")
}
