% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frn_nbm_bl.R
\name{get_frn_nbm_bl}
\alias{get_frn_nbm_bl}
\title{Load part of NBM at Census Block from CORI s3 bucket}
\usage{
get_frn_nbm_bl(frn)
}
\arguments{
\item{frn}{a string of 10 numbers matching FCC's FRN, default is "all"}
}
\value{
a data frame
}
\description{
Get all the data related to a FRN.
A row in this data represent a census block (2010 vintage).
Use \code{get_fcc_dictionary("nbm_block")} to get a description of the date.
A FRN is a 10 number strings, ie "0007435902" can also be used to be more specific.
}
\details{
FCC Broadband Data Collection
}
\examples{
\dontrun{
skymesh <- get_frn_nbm_bl("0027136753")
}
}
