% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_available.R
\name{get_nbm_available}
\alias{get_nbm_available}
\title{Get release available in FCC NBM}
\usage{
get_nbm_available(
  get_data_url = paste0("https://broadbandmap.fcc.gov/nbm/map/",
    "api/national_map_process/nbm_get_data_download/"),
  user_agent = the$user_agent
)
}
\arguments{
\item{get_data_url}{a string providing NBM filing API.}

\item{user_agent}{a string set up by default}
}
\value{
A data frame.
}
\description{
NBM's API:

\if{html}{\out{<div class="sourceCode">}}\preformatted{paste0("https://broadbandmap.fcc.gov/nbm/",
       "map/api/national_map_process/nbm_get_data_download/")
}\if{html}{\out{</div>}}
}
\examples{
nbm_data <- get_nbm_available()
}
