% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_bl.R
\name{get_nbm_bl}
\alias{get_nbm_bl}
\title{Load part of NBM at Census Block from CORI s3 bucket}
\usage{
get_nbm_bl(geoid_co)
}
\arguments{
\item{geoid_co}{a string of 5-digit numbers}
}
\value{
a data frame
}
\description{
Get all the data related to a states or county.
}
\details{
A row in this data represent a census block (2020 vintage).
Use \code{get_fcc_dictionary("nbm_block")} to get a description of the date.

Data Source: FCC Broadband Data Collection
}
\examples{
nbm_bl <- get_nbm_bl(geoid_co = "47051")
}
