% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbm_county.R
\name{get_nbm_county}
\alias{get_nbm_county}
\title{Load part of NBM counts for givent Census County from CORI S3 bucket}
\usage{
get_nbm_county(geoid_co, release = "latest")
}
\arguments{
\item{geoid_co}{a string of 5-digit numbers}

\item{release}{a string with value "D23", "J24", "D24" (respectively targeting releases from December 2023, June 2024, December 2024)}
}
\value{
a data frame
}
\description{
Get all the data related to a county.
}
\details{
A row in this data represent a Census County (2020 vintage).

Data Source: FCC Broadband Data Collection
}
\examples{
\dontrun{
  nbm_bl <- get_nbm_county(geoid_co = "47051", release = "D24")
}
}
\keyword{internal}
